+++ src/OpenRiaServices.Server/Framework/Data/DomainService.cs
@@ -347,6 +347,12 @@ namespace OpenRiaServices.Server
                 // One or more results were returned. If the result is enumerable, compose
                 // any specified query operators, otherwise just return the singleton instance.
                 enumerableResult = result as IEnumerable;
+#if NET
+                if (enumerableResult == null && result is IAsyncEnumerable<T> asyncEnumerableResult)
+                {
+                    enumerableResult = await EnumerateAsyncEnumerable(asyncEnumerableResult, DomainService.DefaultEstimatedQueryResultCount, cancellationToken).ConfigureAwait(false);
+                }
+#endif
                 if (enumerableResult != null)
                 {
                     // If there are additional filtering, sorting and paging parameters to apply
@@ -1254,17 +1260,6 @@ namespace OpenRiaServices.Server
             {
                 return EnumerateAsyncEnumerable(asyncEnumerable, estimatedResultCount, cancellationToken);
             }
-
-            static async ValueTask<IReadOnlyCollection<T>> EnumerateAsyncEnumerable(IAsyncEnumerable<T> asyncEnumerable, int estimatedResultCount, CancellationToken cancellationToken)
-            {
-                List<T> result = new List<T>(capacity: estimatedResultCount);
-                await foreach (var item in asyncEnumerable.WithCancellation(cancellationToken).ConfigureAwait(false))
-                {
-                    result.Add(item);
-                }
-
-                return result;
-            }
 #endif
 
             var list = new List<T>(estimatedResultCount);
@@ -1274,6 +1269,19 @@ namespace OpenRiaServices.Server
             }
             return new ValueTask<IReadOnlyCollection<T>>(list);
         }
+
+#if NET
+        private static async ValueTask<IReadOnlyCollection<T>> EnumerateAsyncEnumerable<T>(IAsyncEnumerable<T> asyncEnumerable, int estimatedResultCount, CancellationToken cancellationToken)
+        {
+            List<T> result = new List<T>(capacity: estimatedResultCount);
+            await foreach (var item in asyncEnumerable.WithCancellation(cancellationToken).ConfigureAwait(false))
+            {
+                result.Add(item);
+            }
+
+            return result;
+        }
+#endif
 
         #region Nested Types